% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.my_marginaleffects}
\alias{knit_print.my_marginaleffects}
\title{Print a my_marginaleffects object in knitr}
\usage{
\method{knit_print}{my_marginaleffects}(x, ...)
}
\value{
A string with class \code{knit_asis} to be printed in Rmarkdown or Quarto documents.
}
\description{
Print a my_marginaleffects object in knitr
}
\keyword{internal}
